requirements.txt
typer==0.3.2
colorama==0.4.4
shellingham==1.4.0
pytest==6.2.4
psycopg2
sqlalchemy
python-dotenv
bcrypt